[Media] Move getFormattedColumn method into the class that uses it
This moves the column formatter for the media module table into the
class that uses it, rather than using a naked function.
